requirements.txt
pycryptodome
numpy
Pillow
matplotlib